requirements.txt
langchain==0.0.304
python-dotenv==1.0.0
streamlit==1.26.0
unstructured==0.9.2
#tiktoken==0.5.1
#faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
openai==0.28.0
python-magic==0.4.27